move testing helper to testing folder
