+++ app/react/V2/testing/index.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line camelcase
+export { LEGACY_createStore } from './testingHelpers';
