+++ app/react/V2/testing/testingHelpers.ts
@@ -1,13 +1,8 @@
 /* eslint-disable camelcase */
 import configureStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import { createStore } from 'jotai';
 import { fromJS } from 'immutable';
-import { merge } from 'lodash';
-import createMockStore from 'redux-mock-store';
 import { IStore } from 'app/istore';
-import { ClientSettings } from 'app/apiResponseTypes';
-import { settingsAtom } from '../atoms';
 
 const defaultState = {
   locale: 'en',
@@ -27,15 +22,4 @@ const middlewares = [thunk];
 const LEGACY_createStore = (state?: Partial<IStore>) =>
   configureStore<object>(middlewares)(() => ({ ...defaultState, ...state }));
 
-const reduxStore = createMockStore([thunk])(() => ({
-  locale: 'en',
-  inlineEdit: fromJS({ inlineEdit: true }),
-  translations: fromJS([
-    {
-      locale: 'en',
-      contexts: [],
-    },
-  ]),
-}));
-
-export { LEGACY_createStore, reduxStore };
+export { LEGACY_createStore };
