+++ app/react/stories/ConfirmationModal.stories.tsx
@@ -4,7 +4,7 @@ import { fn } from '@storybook/test';
 import { action } from '@storybook/addon-actions';
 import { Provider } from 'react-redux';
 import { ConfirmationModal } from 'app/V2/Components/UI/ConfirmationModal';
-import { LEGACY_createStore as createStore } from 'V2/shared/testingHelpers';
+import { LEGACY_createStore as createStore } from 'V2/testing';
 import { Translate } from 'app/I18N';
 
 const meta: Meta<typeof ConfirmationModal> = {
