+++ app/react/stories/Forms/EnableButtonCheckbox.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Meta, StoryObj } from '@storybook/react';
 import { EnableButtonCheckbox } from 'app/V2/Components/Forms';
-import { LEGACY_createStore as createStore } from 'V2/shared/testingHelpers';
+import { LEGACY_createStore as createStore } from 'V2/testing';
 import { Provider } from 'react-redux';
 
 const meta: Meta<typeof EnableButtonCheckbox> = {
