+++ app/react/stories/Forms/FileDropzone.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Meta, StoryObj } from '@storybook/react';
 import { FileDropzone } from 'app/V2/Components/Forms';
-import { LEGACY_createStore as createStore } from 'V2/shared/testingHelpers';
+import { LEGACY_createStore as createStore } from 'V2/testing';
 import { Provider } from 'react-redux';
 
 const meta: Meta<typeof FileDropzone> = {
