+++ app/react/stories/Forms/Geolocation.stories.tsx
@@ -3,7 +3,7 @@ import { Meta, StoryObj } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { fn } from '@storybook/test';
 import { Geolocation } from 'app/V2/Components/Forms';
-import { LEGACY_createStore as createStore } from 'V2/shared/testingHelpers';
+import { LEGACY_createStore as createStore } from 'V2/testing';
 import { Provider } from 'react-redux';
 
 const meta: Meta<typeof Geolocation> = {
