+++ app/react/stories/Forms/MultiselectList.stories.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { Meta, StoryObj } from '@storybook/react';
 import { Provider } from 'react-redux';
 import { MultiselectList } from 'V2/Components/Forms';
-import { LEGACY_createStore as createStore } from 'V2/shared/testingHelpers';
+import { LEGACY_createStore as createStore } from 'V2/testing';
 
 const meta: Meta<typeof MultiselectList> = {
   title: 'Forms/MultiselectList',
