+++ app/react/stories/MediaPlayer.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
 import { Provider } from 'react-redux';
-import { LEGACY_createStore as createStore } from 'V2/shared/testingHelpers';
+import { LEGACY_createStore as createStore } from 'V2/testing';
 import { MediaPlayer } from 'V2/Components/UI';
 
 const meta: Meta<typeof MediaPlayer> = {
