+++ app/react/stories/Table.stories.tsx
@@ -5,7 +5,7 @@ import { action } from '@storybook/addon-actions';
 import { Cell, createColumnHelper, SortingState } from '@tanstack/react-table';
 import { Provider } from 'react-redux';
 import { Button, Table } from 'V2/Components/UI';
-import { LEGACY_createStore as createStore } from 'V2/shared/testingHelpers';
+import { LEGACY_createStore as createStore } from 'V2/testing';
 import { BasicData, DataWithGroups, basicData, dataWithGroups } from './table/fixtures';
 
 type StoryProps = {
